# web/app/api/auth/register/route.ts
import { AxiosError } from 'axios';
import { NextResponse } from 'next/server';

import { registerApiV1AuthRegisterPost, type TokenResponse } from '@/client';
import { getServerClient } from '@/lib/api/client';

interface RegisterRequestBody {
  email: string;
  name: string;
  password: string;
}

export async function POST(request: Request) {
  try {
    const body: RegisterRequestBody = await request.json();

    const { data, error } = await registerApiV1AuthRegisterPost({
      client: getServerClient(),
      body: {
        email: body.email,
        name: body.name,
        password: body.password,
      },
    });

    if (error) {
      const axiosError = error as AxiosError;
      const status = axiosError.response?.status ?? 500;

      if (status === 400) {
        return NextResponse.json(
          { error: 'emailAlreadyRegistered' },
          { status: 400 }
        );
      }

      if (status === 422) {
        return NextResponse.json({ error: 'validationError' }, { status: 422 });
      }

      return NextResponse.json({ error: 'registrationFailed' }, { status });
    }

    if (!data) {
      return NextResponse.json(
        { error: 'registrationFailed' },
        { status: 500 }
      );
    }

    const tokens = data as TokenResponse;

    return NextResponse.json(tokens, { status: 201 });
  } catch {
    return NextResponse.json({ error: 'registrationFailed' }, { status: 500 });
  }
}