# web/openapi.json
{"openapi": "3.1.0", "info": {"title": "ragxuary", "description": "RAG-native documentation tool API", "version": "0.1.0"}, "paths": {"/api/v1/health": {"get": {"tags": ["health"], "summary": "Health Check", "description": "Health check endpoint.\n\nReturns the health status of the application and database connection.", "operationId": "health_check_api_v1_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HealthResponse"}}}}}}}, "/api/v1/auth/register": {"post": {"tags": ["auth"], "summary": "Register", "description": "Register a new user.\n\nArgs:\n    request: Registration request containing email, name, and password.\n    db: Database session.\n\nReturns:\n    Access and refresh tokens.\n\nRaises:\n    HTTPException: If email is already registered.", "operationId": "register_api_v1_auth_register_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RegisterRequest"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/login": {"post": {"tags": ["auth"], "summary": "Login", "description": "Login with email and password.\n\nArgs:\n    request: Login request containing email and password.\n    db: Database session.\n\nReturns:\n    Access and refresh tokens.\n\nRaises:\n    HTTPException: If credentials are invalid.", "operationId": "login_api_v1_auth_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/logout": {"post": {"tags": ["auth"], "summary": "Logout", "description": "Logout and invalidate the current access token.\n\nArgs:\n    credentials: HTTP Bearer credentials containing the JWT token.", "operationId": "logout_api_v1_auth_logout_post", "responses": {"204": {"description": "Successful Response"}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/auth/refresh": {"post": {"tags": ["auth"], "summary": "Refresh", "description": "Refresh access token using a refresh token.\n\nArgs:\n    request: Refresh token request.\n    db: Database session.\n\nReturns:\n    New access and refresh tokens.\n\nRaises:\n    HTTPException: If refresh token is invalid.", "operationId": "refresh_api_v1_auth_refresh_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RefreshTokenRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/me": {"get": {"tags": ["auth"], "summary": "Get Current User Info", "description": "Get current user information.\n\nArgs:\n    current_user: The authenticated user.\n\nReturns:\n    User information.", "operationId": "get_current_user_info_api_v1_auth_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserRead"}}}}}, "security": [{"HTTPBearer": []}]}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "HealthResponse": {"properties": {"status": {"type": "string", "title": "Status"}, "database": {"type": "string", "title": "Database"}}, "type": "object", "required": ["status", "database"], "title": "HealthResponse", "description": "Health check response schema."}, "LoginRequest": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["email", "password"], "title": "LoginRequest", "description": "Schema for login request."}, "RefreshTokenRequest": {"properties": {"refresh_token": {"type": "string", "title": "Refresh Token"}}, "type": "object", "required": ["refresh_token"], "title": "RefreshTokenRequest", "description": "Schema for token refresh request."}, "RegisterRequest": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "name": {"type": "string", "title": "Name"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["email", "name", "password"], "title": "RegisterRequest", "description": "Schema for user registration request."}, "TokenResponse": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "refresh_token": {"type": "string", "title": "Refresh Token"}, "token_type": {"type": "string", "title": "Token Type", "default": "bearer"}}, "type": "object", "required": ["access_token", "refresh_token"], "title": "TokenResponse", "description": "Schema for token response."}, "UserRead": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "name": {"type": "string", "title": "Name"}, "id": {"type": "string", "format": "uuid", "title": "Id"}, "auth_provider": {"type": "string", "title": "Auth Provider"}, "is_active": {"type": "boolean", "title": "Is Active"}, "is_admin": {"type": "boolean", "title": "Is Admin"}, "api_limit": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Api Limit"}, "preferred_locale": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Preferred Locale"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["email", "name", "id", "auth_provider", "is_active", "is_admin", "api_limit", "preferred_locale", "created_at", "updated_at"], "title": "UserRead", "description": "Schema for reading a user."}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"HTTPBearer": {"type": "http", "scheme": "bearer"}}}}